Initial commit of the C++ API
This commit adds the initial version of the C++ API.

This commits contains the source code, the Makefile, required
external libraries, an example and the Doxygen configuration file
for generating the documentation.

In the README and in the Makefile there are references to test code
which is not part of this commit and that will be added later.